Move allocation of analyzer_hash to file_mgr
